+++ src/client/bin/main.rs
@@ -1,10 +1,10 @@
 extern crate cartel;
 
 use cartel::client::cli::cli_app;
-use cartel::{teprint, texit};
+use cartel::{teprinterr, texiterr};
 
 fn main() {
     if let Err(e) = cli_app() {
-        texit!(e);
+        texiterr!(e);
     }
 }